+++ chaincode/src/test/java/de/upb/cs/uc4/chaincode/mock/MockKeyValue.java
@@ -1,10 +1,12 @@
 package de.upb.cs.uc4.chaincode.mock;
 
-import de.upb.cs.uc4.chaincode.GsonWrapper;
+import de.upb.cs.uc4.chaincode.util.GsonWrapper;
 import de.upb.cs.uc4.chaincode.model.MatriculationData;
 import org.hyperledger.fabric.shim.ledger.KeyValue;
 
 import java.util.Objects;
+
+import static java.nio.charset.StandardCharsets.UTF_8;
 
 final public class MockKeyValue implements KeyValue {
 
@@ -24,7 +26,7 @@ final public class MockKeyValue implements KeyValue {
 
     @Override
     public String getStringValue() {
-        return this.value.toString();
+        return new String(this.value, UTF_8);
     }
 
     @Override
@@ -41,10 +43,9 @@ final public class MockKeyValue implements KeyValue {
             return false;
         }
         MockKeyValue other = (MockKeyValue) o;
-        GsonWrapper gson = new GsonWrapper();
         return Objects.equals(this.key, other.key) &&
                 Objects.equals(
-                        gson.fromJson(this.value.toString(), MatriculationData.class),
-                        gson.fromJson(other.value.toString(), MatriculationData.class));
+                        GsonWrapper.fromJson(new String(this.value, UTF_8), MatriculationData.class),
+                        GsonWrapper.fromJson(new String(other.value, UTF_8), MatriculationData.class));
     }
 }